Pivot to github actions instead of travis (#66)
* Pivot to github actions instead of travis

The translated documentation builds and upload time can consistently
take well over 1 hour. The travis jobs have a hard limit of 50min and
this is similar in other CI systems like azure-pipelines which is used
elsewhere in qiskit. According to the docs for github actions the job
timeout is 6 hours [1] which should be more than enough for the docs
builds. This commit adds an initial workflow for publish docs and
removes the travis job config so we will rely on a post merge action to
build and publish the documentation moving forward.

[1] https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#usage-limits

* Update .github/workflows/main.yml