Add stable builds for translations on tag events (#523)
* Add stable builds for translations on tag events

This commit adds stable builds on github tag events. Now that we have
stable builds hosted on qiskit.org/documentation (since
Qiskit/qiskit#1261) we can add stable versions of the translations
corresponding to the same releases. This adds a new bash script and gha
ci job to build the documentation on tag events that will have the
proper content prefix set and upload the documentation to
documentation/stable/$VERSION/locale/$LANG (instead of the usual
docuemntation/locale/$LANG). This will correspond with the hosted stable
version of the documentation for any given version and enable users to
read translated documentation for a given minor release corresponding to
the tag.