Update job definitions to do a full git checkout (#524)
CI builds are failing because the docs are not doing a full checkout.
Since the introduction of stable versions of the docs we need to have
the git tags fetched locally so we can determine which version string to
to use for the current builds (relative to the stable builds). However
the github actions configuration was doing a shallow clone which
excludes the tags. This causes the job to fail when git describe was
used to determine the working version. This commit fixes this by setting
the config flag in the job definitions to do a full clone which will
have the necessary tags.