Update LOCALIZATION_CONTRIBUTORS (#567)
I'm a senior high school in Taiwan and has a great interest in quantum computing. Recently I also competed in the qiskit quantum challenge 2021 and finished the challenge with 16th best score. I would be my honor to contribute to the community and help to translate the document to traditional Chinese. Thanks in advance.

Co-authored-by: Helena Zhang <helena@astbb.com>
skip ci